# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SaaS backend with Keycloak-issued JWT auth.\nAll protected endpoints require a Bearer token obtained from Keycloak.",
        "title": "Lightweight SaaS Backend API",
        "contact": {},
        "version": "0.4.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/workspaces/{workspace_id}/audit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ProjectKeyAuth": []
                    }
                ],
                "description": "Returns the durable audit trail for this workspace, newest first.\n\n**Required scope (project credentials):** `audit:read`.\nOperators are authorized by the realm `admin` role instead.\n\nThe trail records ATTEMPTS TO CHANGE STATE by an identified actor —\nwho did what, to which resource, whether it worked, and on which\nrequest. It is not a request log: reads are absent, and so is\ntraffic that never got past authentication.\n\n**Pagination is cursor-based.** Pass the `next_cursor` from the\nprevious page; its ABSENCE means there is no more history. Offset\npagination is not offered: this table only ever grows at the head,\nso an offset shifts under a client between pages.\n\n**This endpoint never returns secret material.** No password, no\ncredential secret or hash, no connection secret, no provider token,\nand no request body — regardless of who is asking.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "List a workspace's audit history",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ws_3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                        "description": "workspace id",
                        "name": "workspace_id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "exact event type, e.g. user.created",
                        "name": "event",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "operator",
                            "project"
                        ],
                        "type": "string",
                        "description": "operator or project",
                        "name": "actor_type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "success",
                            "failure"
                        ],
                        "type": "string",
                        "description": "success or failure",
                        "name": "outcome",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-08-01T00:00:00Z",
                        "description": "inclusive lower bound, RFC 3339",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-08-31T23:59:59Z",
                        "description": "inclusive upper bound, RFC 3339",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "opaque cursor from the previous page",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size, 1-200 (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auditlog.ListEventsResponse"
                        }
                    },
                    "400": {
                        "description": "invalid_request, invalid_workspace_id",
                        "schema": {
                            "$ref": "#/definitions/auditlog.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "credential_invalid",
                        "schema": {
                            "$ref": "#/definitions/auditlog.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden, workspace_mismatch, insufficient_scope",
                        "schema": {
                            "$ref": "#/definitions/auditlog.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "audit_unavailable",
                        "schema": {
                            "$ref": "#/definitions/auditlog.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auditlog.ActorResponse": {
            "type": "object",
            "properties": {
                "credential_id": {
                    "type": "string",
                    "example": "key_9b2f4c1a-1111-4222-8333-444455556666"
                },
                "email": {
                    "type": "string",
                    "example": "ada@example.com"
                },
                "project_id": {
                    "type": "string",
                    "example": "prj_7c9e6679-7425-40de-944b-e07fc1f90ae7"
                },
                "subject": {
                    "type": "string",
                    "example": "9c1e6679-7425-40de-944b-e07fc1f90ae7"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "operator",
                        "project"
                    ],
                    "example": "operator"
                }
            }
        },
        "auditlog.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "field": {
                    "type": "string",
                    "example": "cursor"
                },
                "message": {
                    "type": "string",
                    "example": "Request is invalid"
                },
                "request_id": {
                    "type": "string",
                    "example": "3f2504e0-4f89-41d3-9a0c-0305e82c3301"
                }
            }
        },
        "auditlog.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/auditlog.ErrorBody"
                }
            }
        },
        "auditlog.EventResponse": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/auditlog.ActorResponse"
                },
                "event": {
                    "type": "string",
                    "example": "project_credential.revoked"
                },
                "id": {
                    "type": "string",
                    "example": "evt_3f2504e0-4f89-41d3-9a0c-0305e82c3301"
                },
                "metadata": {
                    "description": "Metadata is per-event detail, allowlisted per event type at write time.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "occurred_at": {
                    "type": "string",
                    "example": "2026-08-10T14:03:11Z"
                },
                "outcome": {
                    "description": "Outcome is \"success\" or \"failure\".",
                    "type": "string",
                    "enum": [
                        "success",
                        "failure"
                    ],
                    "example": "success"
                },
                "reason_code": {
                    "description": "ReasonCode is present only on a failure, and is drawn from the closed /v1\nerror vocabulary. It is never an upstream error message: the real cause\nis in the log line for this request_id.",
                    "type": "string",
                    "example": "provider_unavailable"
                },
                "request_id": {
                    "type": "string",
                    "example": "3f2504e0-4f89-41d3-9a0c-0305e82c3301"
                },
                "resource": {
                    "$ref": "#/definitions/auditlog.ResourceResponse"
                }
            }
        },
        "auditlog.ListEventsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auditlog.EventResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/auditlog.PaginationInfo"
                }
            }
        },
        "auditlog.PaginationInfo": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Count is the number of items IN THIS PAGE, never a total. A total would\nmean a second COUNT(*) over an append-heavy table on every request, to\nanswer a question no client here needs.",
                    "type": "integer",
                    "example": 50
                },
                "limit": {
                    "description": "Limit is the page size actually applied.",
                    "type": "integer",
                    "example": 50
                },
                "next_cursor": {
                    "description": "NextCursor is absent on the last page.\n\nIts ABSENCE is the end-of-history signal, not an empty page: a client\nloops while this is present, so a correct client makes exactly as many\nrequests as there are pages.",
                    "type": "string",
                    "example": "MTc1NDg0MTM5MTAwMDAwMDAwMC4z..."
                }
            }
        },
        "auditlog.ResourceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "key_9b2f4c1a-1111-4222-8333-444455556666"
                },
                "type": {
                    "type": "string",
                    "example": "project_credential"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "OPERATOR authentication. Type \"Bearer\" followed by a Keycloak-issued access token. The token must carry the realm `admin` role, which is re-checked against the provider on every request.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ProjectKeyAuth": {
            "description": "PROJECT authentication (machine-to-machine). Type \"Bearer\" followed by a project credential, e.g. `Bearer lw_sk_\u003clookup\u003e_\u003csecret\u003e`. The credential is bound to exactly one workspace and carries an explicit scope set; the scope each operation requires is stated in its description. Control-plane operations (workspaces, connections, projects) are operator-only and answer `operator_only` to any credential. Swagger 2.0 cannot express per-operation scopes for an apiKey scheme, so they are documented in prose rather than declared — see docs/PROJECTS.md.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}